Address CR comments regarding RandomWhateverTests and moved the method
